Add initial boilerplating for creating gem packaging
Add a task to generate a license

Add a task similar to `bundle gem` but generating file from local erb templates.  Functionality will be added shorly to update the files as well as create them, giving us the ability to make a change to the README in every repo at once.